requirements.txt
asgiref==3.7.2
bleach==6.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cloudinary==1.39.1
cryptography==41.0.4
defusedxml==0.7.1
dj-database-url==2.1.0
dj-rest-auth==4.0.0
Django==4.2.5
django-allauth==0.52.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
django-jazzmin==2.6.0
django-summernote==0.8.20.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.7.0
oauthlib==3.2.2
packaging==24.0
Pillow==10.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
stripe==8.7.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.5
webencodings==0.5.1
whitenoise==6.6.0
zipp==3.16.2
